Change QR to Preference Item, and Make App only request Camera permissions when needed in the QRScanner Activity
