+++ PlexPy Remote/src/main/java/com/williamcomartin/plexpyremote/AppBaseActivity.java
@@ -22,19 +22,7 @@ public class AppBaseActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-        requestCameraPermission();
-
         Iconify.with(new FontAwesomeModule()).with(new MaterialModule());
-
         setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
-
-    }
-
-    public void requestCameraPermission() {
-        if (ContextCompat.checkSelfPermission(this,
-                Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.CAMERA}, MY_PERMISSIONS_REQUEST_CAMERA);
-        }
     }
 }
