--- PlexPy Remote/src/main/res/menu/menu_main.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/readQRButton"
-        android:icon="@drawable/qrcode"
-        android:orderInCategory="20"
-        app:showAsAction="ifRoom"
-        android:title="QR Reader">
-    </item>
-
-</menu>
