+++ PlexPy Remote/src/main/res/values/attrs.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="EditTextPreferenceWithDisplay">
+        <attr name="defaultSummary" format="string" />
+    </declare-styleable>
+</resources>
